% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_param_sweep_modelsel.R
\name{aggregate_modelsel_results}
\alias{aggregate_modelsel_results}
\title{Aggregate Model Selection Results}
\usage{
aggregate_modelsel_results(results_df)
}
\arguments{
\item{results_df}{Raw results data frame from sweep.}
}
\value{
Aggregated data frame.
}
\description{
Aggregate Model Selection Results
}
\keyword{internal}
