% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{align_bootstrap_loadings}
\alias{align_bootstrap_loadings}
\title{Align Bootstrap Loadings to Reference Using Procrustes}
\usage{
align_bootstrap_loadings(boot_W, reference_W, boot_z)
}
\arguments{
\item{boot_W}{List of K loading matrices from bootstrap sample}

\item{reference_W}{List of K loading matrices from reference fit}

\item{boot_z}{Bootstrap cluster assignments (for label switching detection)}
}
\value{
List of K aligned loading matrices
}
\description{
Internal function that aligns bootstrap loading matrices to reference
loadings using orthogonal Procrustes rotation and handles cluster
label switching.
}
\keyword{internal}
