% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_filters.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply Filter to EMG Data}
\usage{
apply_filter(data, bf, zero_phase)
}
\arguments{
\item{data}{EMG data (vector, matrix, or list)}

\item{bf}{Filter object from signal package}

\item{zero_phase}{Whether to use zero-phase filtering}
}
\value{
Filtered data
}
\description{
Internal function to apply a filter to various data structures.
}
\keyword{internal}
