% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{calc_reconstruction_error_twostep}
\alias{calc_reconstruction_error_twostep}
\title{Helper function to compute SSE for Two-step baseline}
\usage{
calc_reconstruction_error_twostep(data_list, two_step_res)
}
\arguments{
\item{data_list}{List of data matrices}

\item{two_step_res}{Result from fit_two_step_baseline}
}
\value{
SSE value
}
\description{
Helper function to compute SSE for Two-step baseline
}
\keyword{internal}
