% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_utils.R
\name{check_future_available}
\alias{check_future_available}
\title{Check if Future Packages are Available}
\usage{
check_future_available(batchtools = FALSE)
}
\arguments{
\item{batchtools}{Logical. Also check for future.batchtools.}
}
\value{
Logical. TRUE if packages are available.
}
\description{
Checks whether the future ecosystem packages are installed.
}
