% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_utils.R
\name{check_pbs_job_status}
\alias{check_pbs_job_status}
\title{Check PBS Job Status}
\usage{
check_pbs_job_status(job_id)
}
\arguments{
\item{job_id}{Character. PBS job ID.}
}
\value{
List with job status information.
}
\description{
Checks the status of a PBS job.
}
\examples{
\dontrun{
status <- check_pbs_job_status("12345")
}

}
