% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_BIC}
\alias{compute_BIC}
\title{Compute Bayesian Information Criterion (BIC)}
\usage{
compute_BIC(loglik, K, r, M, N_obs, model_type)
}
\arguments{
\item{loglik}{Numeric scalar, the log-likelihood value}

\item{K}{Integer, number of clusters}

\item{r}{Integer, number of factors/synergies}

\item{M}{Integer, number of variables (muscles)}

\item{N_obs}{Integer, number of observations (total time points across subjects)}

\item{model_type}{Character, either "MFA" or "MPCA"}
}
\value{
A numeric scalar representing the BIC value (lower is better)
}
\description{
Unified interface for computing BIC for both MFA and MPCA models.
}
\details{
BIC = -2 * loglik + p * log(N_obs), where p is the number of parameters.
The parameter count differs between MFA and MPCA due to different model
structures.
}
\examples{
\dontrun{
# Compute BIC for an MFA model
bic <- compute_BIC(loglik, K = 2, r = 3, M = 8, N_obs = 1000, model_type = "MFA")
}

}
\seealso{
\code{\link{compute_BIC_mfa}}, \code{\link{compute_BIC_mpca}}
}
