% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_held_out_loglik}
\alias{compute_held_out_loglik}
\title{Compute Log-Likelihood on Held-Out Data}
\usage{
compute_held_out_loglik(test_data, fit, model_type)
}
\arguments{
\item{test_data}{List of data matrices for held-out subjects.}

\item{fit}{Fitted model object from mfa_em_fit or mpca_em_fit.}

\item{model_type}{Either "MFA" or "MPCA".}
}
\value{
Total log-likelihood of test data.
}
\description{
Internal function to compute log-likelihood of held-out subjects
given a fitted mixture model.
}
\keyword{internal}
