% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_ICL}
\alias{compute_ICL}
\title{Compute ICL for a Mixture Model (Unified Interface)}
\usage{
compute_ICL(loglik, K, r, M, N_obs, resp, model_type)
}
\arguments{
\item{loglik}{The total log-likelihood from the fitted model.}

\item{K}{Number of clusters.}

\item{r}{Factor/principal component dimension.}

\item{M}{Observed dimension (channels).}

\item{N_obs}{Number of observations.}

\item{resp}{An (N x K) matrix of responsibilities.}

\item{model_type}{Either "MFA" or "MPCA".}
}
\value{
A numeric ICL value.
}
\description{
Unified interface for computing ICL (Integrated Complete-data Likelihood)
for either MFA or MPCA models.
}
\seealso{
\code{\link{compute_ICL_mfa}}, \code{\link{compute_ICL_mpca}}
}
