% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_param_sweep_modelsel.R
\name{compute_modelsel_win_rates}
\alias{compute_modelsel_win_rates}
\title{Compute Win Rates for Model Selection}
\usage{
compute_modelsel_win_rates(aggregated_df)
}
\arguments{
\item{aggregated_df}{Aggregated results data frame.}
}
\value{
Data frame with win rates.
}
\description{
Compute Win Rates for Model Selection
}
\keyword{internal}
