% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_envelope.R
\name{envelope_lowpass}
\alias{envelope_lowpass}
\title{Low-pass filter envelope extraction}
\usage{
envelope_lowpass(x, sampling_rate, cutoff, filter_order, zero_phase)
}
\description{
Low-pass filter envelope extraction
}
\keyword{internal}
