% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_envelope.R
\name{envelope_moving_average}
\alias{envelope_moving_average}
\title{Moving average envelope extraction}
\usage{
envelope_moving_average(x, sampling_rate, window_ms)
}
\description{
Moving average envelope extraction
}
\keyword{internal}
