% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_envelope.R
\name{extract_envelope_vector}
\alias{extract_envelope_vector}
\title{Extract envelope from a single vector}
\usage{
extract_envelope_vector(
  x,
  sampling_rate,
  method,
  cutoff,
  window_ms,
  filter_order,
  zero_phase
)
}
\description{
Extract envelope from a single vector
}
\keyword{internal}
