% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_normalize.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalize a single vector}
\usage{
normalize_vector(x, method, reference = NULL)
}
\description{
Normalize a single vector
}
\keyword{internal}
