% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{print.bootstrap_ci}
\alias{print.bootstrap_ci}
\title{Print Summary of Bootstrap CI Results}
\usage{
\method{print}{bootstrap_ci}(x, cluster = 1, ...)
}
\arguments{
\item{x}{Result from bootstrap_loading_ci}

\item{cluster}{Which cluster to summarize (default: 1)}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Summary of Bootstrap CI Results
}
