% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{print.calibration_result}
\alias{print.calibration_result}
\title{Print Summary of Calibration Results}
\usage{
\method{print}{calibration_result}(x, ...)
}
\arguments{
\item{x}{A calibration_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Summary of Calibration Results
}
