% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_selection_mixture.R
\name{print.mixture_r_selection}
\alias{print.mixture_r_selection}
\title{Print Method for mixture_r_selection Objects}
\usage{
\method{print}{mixture_r_selection}(x, ...)
}
\arguments{
\item{x}{A mixture_r_selection object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for mixture_r_selection Objects
}
