% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_main.R
\name{print.preprocessed_emg}
\alias{print.preprocessed_emg}
\title{Print Method for preprocessed_emg Objects}
\usage{
\method{print}{preprocessed_emg}(x, ...)
}
\arguments{
\item{x}{A preprocessed_emg object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for preprocessed_emg Objects
}
