% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{reconstruct_singlePCA}
\alias{reconstruct_singlePCA}
\title{Reconstruct Data Using Single PCA Model}
\usage{
reconstruct_singlePCA(X, pca_res)
}
\arguments{
\item{X}{Numeric matrix (T x M) of data for one subject.}

\item{pca_res}{PCA result containing P (loadings) and mu (mean).}
}
\value{
A matrix of reconstructed data (T x M).
}
\description{
Internal helper to reconstruct data matrix using PCA parameters.
}
\keyword{internal}
