% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_filters.R
\name{remove_dc_offset_vector}
\alias{remove_dc_offset_vector}
\title{Remove DC offset from a single vector}
\usage{
remove_dc_offset_vector(x, method, poly_order)
}
\description{
Remove DC offset from a single vector
}
\keyword{internal}
