% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-templates.R
\name{return-templates}
\alias{return-templates}
\title{Return Value Templates}
\description{
Standard return value documentation for synergyMixR functions.
}
\section{MFA Fit Return Value}{

For MFA models (mfa_em_fit), returns a list with:
\describe{
  \item{z}{Integer vector of length N. Hard cluster assignments (1 to K).}
  \item{pi}{Numeric vector of length K. Cluster mixing proportions.}
  \item{mu}{List of length K. Each element is an M-dimensional mean vector.}
  \item{Lambda}{List of length K. Each element is an M x r factor loading matrix.}
  \item{Psi}{List of length K. Each element is an M x M diagonal noise covariance.}
  \item{logLik}{Numeric. Final log-likelihood value.}
  \item{resp}{N x K matrix of cluster responsibilities (soft assignments).}
}
}

\section{MPCA Fit Return Value}{

For MPCA models (mixture_pca_em_fit), returns a list with:
\describe{
  \item{z}{Integer vector of length N. Hard cluster assignments (1 to K).}
  \item{pi}{Numeric vector of length K. Cluster mixing proportions.}
  \item{mu}{List of length K. Each element is an M-dimensional mean vector.}
  \item{W}{List of length K. Each element is an M x r projection matrix.}
  \item{P}{List of length K. Each element is an M x r principal directions matrix.}
  \item{D}{List of length K. Each element is an r x r diagonal scaling matrix.}
  \item{Psi}{List of length K. Each element is an M x M diagonal noise covariance.}
  \item{sigma2}{Numeric vector of length K. Noise variances per cluster.}
  \item{logLik}{Numeric. Final log-likelihood value.}
  \item{resp}{N x K matrix of cluster responsibilities (soft assignments).}
}
}

\keyword{internal}
