% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_comparison.R
\name{simulation_comparison}
\alias{simulation_comparison}
\title{Simulation Comparison Functions}
\description{
Functions for comparing different methods (SingleFA, SinglePCA, MixtureFA, MixturePCA)
on simulated data.
}
