% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_selection_mixture.R
\name{summary.mixture_r_selection}
\alias{summary.mixture_r_selection}
\title{Summary Method for mixture_r_selection Objects}
\usage{
\method{summary}{mixture_r_selection}(object, ...)
}
\arguments{
\item{object}{A mixture_r_selection object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary Method for mixture_r_selection Objects
}
