% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_main.R
\name{summary.preprocessed_emg}
\alias{summary.preprocessed_emg}
\title{Summary Method for preprocessed_emg Objects}
\usage{
\method{summary}{preprocessed_emg}(object, ...)
}
\arguments{
\item{object}{A preprocessed_emg object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for preprocessed_emg Objects
}
